create and implement TabContext and implement it in sidebar.tsx
