+++ src/hooks/useTab.ts
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { TabContext } from "../contexts/TabContext";
+
+const useTab = () => {
+  return useContext(TabContext);
+};
+
+export default useTab;
